Improve UDP implementation to not use a channel
